requirements.txt
Django==5.0.6
psycopg2-binary==2.9.9
django-extensions==3.2.3
ipython==8.26.0
pillow==10.3.0
djangorestframework==3.15.2
factory-boy==3.3.0
shortuuid==1.0.13
